Remove useless call to user model & unique name for image to user ctrl

+++ api/controller/UserCtrl.js
@@ -322,55 +322,38 @@ exports.updateUser = (req, res, next) => {
       .then((users) => {
         this.checkUsersForUnique(req.params.id, users, fields, res);
 
-        UserModel
-        .findById(req.params.id)
-        .then(oldUser => {
-          let image = oldUser.image;
+        let image = nem.getName(fields.name) + "." + process.env.IMG_EXT;
+        if (files.image) nem.setThumbnail("users/" + files.image.newFilename, USERS_THUMB + image);
 
-          if (files.image) {
-            image = nem.getUniqueName(fields.name) + "." + process.env.IMG_EXT;
-            nem.setThumbnail("users/" + files.image.newFilename, USERS_THUMB + image);
-          }
+        if (fields.pass) {
+          this.checkUserPass(fields.pass, res);
 
-          if (fields.pass) {
-            this.checkUserPass(fields.pass, res);
-        
-            bcrypt
-            .hash(fields.pass, 10)
-            .then((hash) => { 
-              let user = this.getUserWithPass(fields.name, fields.email, image, hash, fields.role, fields.updated);
-
-              UserModel
-                .findByIdAndUpdate(req.params.id, { ...user, _id: req.params.id })
-                .then(() => {
-                  if (files.image) { 
-                    fs.unlink(USERS_IMG + files.image.newFilename, () => {
-                      fs.unlink(USERS_THUMB + oldUser.image, () => {})
-                    }) 
-                  }
-                  res.status(200).json({ message: process.env.USER_UPDATED });
-                })
-                .catch(() => res.status(400).json({ message: process.env.USER_NOT_UPDATED }));
-            })
-            .catch(() => res.status(400).json({ message: process.env.USER_NOT_PASS }));
-        
-          } else { 
-            let user = this.getUserNoPass(fields.name, fields.email, image, fields.role, fields.updated);
+          bcrypt
+          .hash(fields.pass, 10)
+          .then((hash) => { 
+            let user = this.getUserWithPass(fields.name, fields.email, image, hash, fields.role, fields.updated);
 
             UserModel
               .findByIdAndUpdate(req.params.id, { ...user, _id: req.params.id })
               .then(() => {
-                if (files.image) { 
-                  fs.unlink(USERS_IMG + files.image.newFilename, () => {
-                    fs.unlink(USERS_THUMB + oldUser.image, () => {})
-                  }) 
-                }
+                if (files.image) fs.unlink(USERS_IMG + files.image.newFilename, () => {});
                 res.status(200).json({ message: process.env.USER_UPDATED });
               })
               .catch(() => res.status(400).json({ message: process.env.USER_NOT_UPDATED }));
-          }
-        })
-        .catch(() => res.status(404).json({ message: process.env.USER_NOT_FOUND }));
+          })
+          .catch(() => res.status(400).json({ message: process.env.USER_NOT_PASS }));
+      
+        } else { 
+          let user = this.getUserNoPass(fields.name, fields.email, image, fields.role, fields.updated);
+
+          UserModel
+            .findByIdAndUpdate(req.params.id, { ...user, _id: req.params.id })
+            .then(() => {
+              if (files.image) fs.unlink(USERS_IMG + files.image.newFilename, () => {});
+              res.status(200).json({ message: process.env.USER_UPDATED });
+            })
+            .catch(() => res.status(400).json({ message: process.env.USER_NOT_UPDATED }));
+        }
       })
       .catch(() => res.status(404).json({ message: process.env.USERS_NOT_FOUND }));
   })
